Moved scores logic & declaration to App.jsx
