+++ src/App.jsx
@@ -1,18 +1,32 @@
-import { useState } from 'react'
+import { useEffect, useState } from 'react'
 import './App.css'
 import WelcomeScreen from './components/screens/WelcomeScreen';
 import PlayersScreen from './components/screens/PlayersScreen';
 import AddRounds from './components/screens/AddRoundsScreen';
 import RoundScreen from './components/screens/RoundScreen';
+import MatchRecapScreen from './components/screens/MatchRecapScreen';
 
 function App() {
   const [currentScreen, setCurrentScreen] = useState(0);
   const [userNicknames, setUserNicknames] = useState([]);
   const [matchId, setMatchId] = useState();
   const [numRounds, setNumRounds] = useState(3);
+  const [scores, setScores] = useState([]);
 
-  // console.log(matchId)
-  // console.log(userNicknames);
+  useEffect(() => {
+    setInitialScores();
+  }, [userNicknames]);
+
+  function setInitialScores() {
+    const initialScores = [];
+    userNicknames.forEach((nickname) => {
+      initialScores.push({
+        "nickname": nickname,
+        "score": 0
+      });
+    });
+    setScores(initialScores);
+  }
 
   const changeScreen = (nextScreen) => setCurrentScreen(nextScreen);
 
@@ -44,6 +58,12 @@ function App() {
         matchId={matchId}
         currentScreen={currentScreen}
         changeScreen={changeScreen}
+        scores={scores}
+        setScores={setScores}
+      />}
+      {currentScreen == 4 && <MatchRecapScreen
+        scores={scores}
+        setCurrentScreen={setCurrentScreen}
       />}
     </>
   )
